requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.0.1+cpu
torchvision==0.15.2+cpu
facenet-pytorch==2.5.2

numpy==1.24.4
scikit-learn==1.3.2
matplotlib==3.7.2
streamlit==1.32.2
Pillow==9.5.0

